+++ test/example/javascript/class.js
@@ -1,0 +1,7 @@
+class Class {
+    method(argument) {}
+    get field() {}
+    set field(value) {}
+}
+
+class SubClass extends SuperClass{}