+++ CHANGELOG.md
@@ -7,8 +7,12 @@ This project tries to adhere to [Semantic Versioning](http://semver.org/).
 ### Added
 - Support ES6 Classes (see [#79](https://github.com/Hirse/brackets-outline-list/issues/79))
 
+### Changed
+- Use htmlparser2 to parse XML, HTML, and SVG
+- Use PostCSS Safe Parser to parse CSS, Less, and SCSS
+
 ### Fixed
-- Various PHP issues, for example multipe inheritance [__@pelatx__](https://github.com/pelatx)
+- Various PHP issues, for example multipe inheritance, thanks to [__@pelatx__](https://github.com/pelatx)
 - Use Class icon for PHP Class definitions
 - Don't break on PHP return types (see [#80](https://github.com/Hirse/brackets-outline-list/issues/80))