+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 This project tries to adhere to [Semantic Versioning](http://semver.org/).
 
 
-## Unreleased
+## 1.0.0 - 2017-03-12
 ### Added
 - Support ES6 Classes (see [#79](https://github.com/Hirse/brackets-outline-list/issues/79))